external style file for SettingsScreen.
