+++ src/app/tabs/SettingsScreen/SettingsStyles.ts
@@ -1,0 +1,53 @@
+import { StyleSheet } from 'react-native';
+
+const settingsStyles = StyleSheet.create({
+    bgContainer: {
+    flex: 1,
+    backgroundColor: '#181818',
+  },
+  container: {
+    flex: 1,
+    backgroundColor: '#1e1f1f',
+    padding: 16,
+    marginLeft: 6,
+    marginRight: 6,
+    borderRadius: 8,
+  },
+  groupsContainer: {
+    flexDirection: 'row',
+    justifyContent: 'center',
+    alignItems: 'center',
+    gap: 8,
+  },
+  elementsSpacing: {
+    marginTop: 10,
+  },
+  labelText: {
+    alignSelf: 'center',
+    color: 'white',
+    fontSize: 20,
+    fontWeight: 'bold',
+    marginBottom: 10,
+  },
+  studentGroups: {
+    flexDirection: 'row',
+    justifyContent: 'center',
+    alignItems: 'center',
+    gap: 16,
+  },
+  notifications: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    gap: 40,
+    flexWrap: 'wrap',
+  },
+  notificationsMid: {
+    flexDirection: 'column',
+    alignSelf: 'center',
+    alignItems: 'center',
+    justifyContent: 'center',
+    gap: 12,
+    flexWrap: 'wrap',}
+});
+
+export default settingsStyles;
